adding new functionality, doing mechanical translation of cuckoo filter decoding

+++ app/src/main/java/com/example/bluetooth_sample/data/EntryDao.kt
@@ -10,7 +10,7 @@ const val MINUTES_IN_WINDOW = 20160
 interface EntryDao {
     // get all entries that are less than 14 days old
     @Query("SELECT * FROM entry WHERE (:currTime) - dongleTime < $MINUTES_IN_WINDOW")
-    fun getAll(currTime: Int): List<Entry>
+    fun getAll(currTime: Long): List<Entry>
 
     // insert new entry
     @Insert
@@ -18,5 +18,5 @@ interface EntryDao {
 
     // delete all entries older than 14 days old
     @Query("DELETE FROM entry WHERE (:currTime) - dongleTime >= $MINUTES_IN_WINDOW")
-    fun delete(currTime: Int)
+    fun delete(currTime: Long)
 }
